move fill container strategy to a separate file because it "grew up"
